✨ feat: SJRA-62 Implement end-to-end partition inserting DAG (#22)

+++ radiant/dags/sql/open_data/init/1000_genomes_create_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS {{ params.starrocks_1000_genomes }} (
     locus_id BIGINT NOT NULL,
-    af DECIMAL(7, 6)
+    af DOUBLE
 )
 ENGINE=OLAP
 DISTRIBUTED BY HASH(locus_id) BUCKETS 10
